# uint256: optimize WriteToArray