@holiman 2024-11-29
Nice work!
I wasn't super-fond of the idea of doing `val.WriteToArray32((*[32]byte)(m.store[offset:]))`, so instead I added a method in the same vein as the binary `PutUint64` - methods: `PutUint256` which just writes it into a slice, and it's up to the caller to ensure that there's sufficient size in the destination. 

WDYT?