@minh-bq 2024-11-29
I just leave a comment. Overall, that function looks good to me.